accessible/tests/mochitest/actions/test_media.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=483573
-->
<head>
  <title>HTML5 audio/video tests</title>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../events.js"></script>
  <script type="application/javascript"
          src="../actions.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../states.js"></script>

  <script type="application/javascript">
    /* global testAction */
    SimpleTest.requestCompleteLog(); // To help diagnose bug 1845221

    async function loadAudioSource() {
      /**
       * Setting the source dynamically and wait for it to load,
       * so we can test the accessibility tree of the control in its ready and
       * stable state.
       *
       * See bug 1484048 comment 25 for discussion on how it switches UI when
       * loading a statically declared source.
       */
      await new Promise(resolve => {
        let el = document.getElementById("audio");
        el.addEventListener("canplaythrough", resolve, {once: true});
        el.src = "../bug461281.ogg";
      });

      await doTest();
    }

    async function doTest() {
      // ////////////////////////////////////////////////////////////////////////
      // test actions of audio controls

      todo(false, "Focus test are disabled until bug 494175 is fixed.");

      var audioElm = getAccessible("audio");
      var playBtn = audioElm.firstChild;
      // var scrubber = playBtn.nextSibling.nextSibling.nextSibling;
      var muteBtn = audioElm.lastChild.previousSibling;

      await testAction({
        id: muteBtn,
        actionName: "press",
        eventSeq: [
     //     [EVENT_FOCUS, muteBtn, () => testStates(muteBtn, STATE_FOCUSED)],
          [EVENT_NAME_CHANGE, muteBtn,
           evt => is(evt.accessible.name, "Unmute",
                     "Wrong name of " + prettyName(evt.accessible) + " on focus")],
        ],
      });
     //   {
     //     id: scrubber,
     //     actionName: "activate",
     //     eventSeq: [
     //       [EVENT_FOCUS, scrubber, () => testStates(scrubber, STATE_FOCUSED)]
     //     ]
     //   },
      await testAction({
        id: playBtn,
        actionName: "press",
        eventSeq: [
     //     [EVENT_FOCUS, playBtn, () => testStates(playBtn, STATE_FOCUSED)],
          [EVENT_NAME_CHANGE, playBtn,
           evt => is(evt.accessible.name, "Pause",
                     "Wrong name of " + prettyName(evt.accessible) + " on focus")],
        ],
      });
      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(loadAudioSource);
  </script>
</head>
<body>

  <a target="_blank" rel="opener"
     title="Expose HTML5 video and audio elements' embedded controls through accessibility APIs"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=483573">Mozilla Bug 483573</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <audio id="audio" controls="true"></audio>

  <div id="eventDump"></div>
</body>
</html>
